Test improvements.
