+++ tests/MongoDB.Driver.Tests/MultipleRegistries/MultipleRegistriesLinqTests.cs
@@ -13,129 +13,158 @@
  * limitations under the License.
  */
 
+using System;
 using System.Linq;
 using FluentAssertions;
 using MongoDB.Bson;
 using MongoDB.Bson.Serialization;
+using MongoDB.Bson.Serialization.Attributes;
+using MongoDB.Bson.Serialization.Conventions;
 using MongoDB.Driver.Core.TestHelpers.XunitExtensions;
 using MongoDB.Driver.Linq;
 using MongoDB.Driver.Tests.Linq.Linq3Implementation;
 using Xunit;
+using static MongoDB.Driver.Tests.MultipleRegistriesTestHelpers;
 
-namespace MongoDB.Driver.Tests
+namespace MongoDB.Driver.Tests;
+
+[Trait("Category", "Integration")]
+public class MultipleRegistriesLinqTests
 {
-    [Trait("Category", "Integration")]
-    public class MultipleRegistriesLinqTests
+    [Fact]
+    public void Linq_Concat_across_different_MongoClients_throws()
     {
-        [Fact]
-        public void Linq_Concat_across_different_MongoClients_throws()
-        {
-            var outerDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("outer");
-            var innerDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("inner");
+        var outerDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("outer");
+        var innerDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("inner");
 
-            var databaseName = DriverTestConfiguration.DatabaseNamespace.DatabaseName;
-            var outerColl = DriverTestConfiguration.CreateMongoClient(c => c.SerializationDomain = outerDomain)
-                .GetDatabase(databaseName).GetCollection<Person>("concat_outer");
-            var innerColl = DriverTestConfiguration.CreateMongoClient(c => c.SerializationDomain = innerDomain)
-                .GetDatabase(databaseName).GetCollection<Person>("concat_inner");
+        var databaseName = DriverTestConfiguration.DatabaseNamespace.DatabaseName;
+        var outerColl = DriverTestConfiguration.CreateMongoClient(c => c.SerializationDomain = outerDomain)
+            .GetDatabase(databaseName).GetCollection<Person>("concat_outer");
+        var innerColl = DriverTestConfiguration.CreateMongoClient(c => c.SerializationDomain = innerDomain)
+            .GetDatabase(databaseName).GetCollection<Person>("concat_inner");
 
-            var queryable = outerColl.AsQueryable().Concat(innerColl.AsQueryable());
+        var queryable = outerColl.AsQueryable().Concat(innerColl.AsQueryable());
 
-            var ex = Assert.Throws<ExpressionNotSupportedException>(() => Linq3TestHelpers.Translate<Person, Person>(queryable));
-            ex.Message.Should().Contain("different MongoClients");
-        }
-
-        [Fact]
-        public void Linq_Inject_translates_filter_under_outer_domain()
-        {
-            RequireServer.Check();
-
-            // Custom domain's CustomStringSerializer rewrites strings with a "test" suffix; the filter value
-            // must be serialized through the domain's registry when the Inject path is translated.
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-            customDomain.RegisterSerializer(new CustomStringSerializer());
-
-            var client = DriverTestConfiguration.CreateMongoClient(c => c.SerializationDomain = customDomain);
-            var collection = client
-                .GetDatabase(DriverTestConfiguration.DatabaseNamespace.DatabaseName)
-                .GetCollection<Person>(DriverTestConfiguration.CollectionNamespace.CollectionName);
-
-            var injectedFilter = Builders<Person>.Filter.Eq(p => p.Name, "Mario");
-            var queryable = collection.AsQueryable().Where(p => injectedFilter.Inject());
-
-            var stages = Linq3TestHelpers.Translate<Person, Person>(queryable);
-            var match = stages.Single(s => s.Contains("$match"));
-            match.ToJson().Should().Contain("Mariotest");
-        }
-
-        [Fact]
-        public void Linq_math_methods_translate_under_custom_domain()
-        {
-            RequireServer.Check();
-
-            // Smoke test: aggregation-expression translators that previously called BsonSerializer.LookupSerializer
-            // (Ceiling, Floor) now route through context.SerializationDomain.
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            var client = DriverTestConfiguration.CreateMongoClient(c => c.SerializationDomain = customDomain);
-            var collection = client
-                .GetDatabase(DriverTestConfiguration.DatabaseNamespace.DatabaseName)
-                .GetCollection<NumericModel>(DriverTestConfiguration.CollectionNamespace.CollectionName);
-
-            var queryable = collection
-                .AsQueryable()
-                .Select(m => new { Ceil = System.Math.Ceiling(m.D), Flr = System.Math.Floor(m.D) });
-
-            var stages = Linq3TestHelpers.Translate<NumericModel, object>(queryable);
-            var project = stages.Single(s => s.Contains("$project"));
-            project.ToJson().Should().Contain("$ceil");
-            project.ToJson().Should().Contain("$floor");
-        }
-
-        [Fact]
-        public void Linq_Pick_with_SortBy_runs_under_custom_domain()
-        {
-            RequireServer.Check();
-
-            // Smoke test: Pick's SortDefinition render path now receives the custom SerializationDomain.
-            // The assertion is that translation completes without reaching the global static registry.
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            var client = DriverTestConfiguration.CreateMongoClient(c => c.SerializationDomain = customDomain);
-            var collection = client
-                .GetDatabase(DriverTestConfiguration.DatabaseNamespace.DatabaseName)
-                .GetCollection<Person>(DriverTestConfiguration.CollectionNamespace.CollectionName);
-
-            var queryable = collection
-                .AsQueryable()
-                .GroupBy(p => p.Name)
-                .Select(g => new { Name = g.Key, YoungestAge = g.BottomN(Builders<Person>.Sort.Descending(p => p.Age), p => p.Age, 1) });
-
-            var stages = Linq3TestHelpers.Translate<Person, object>(queryable);
-            var group = stages.Single(s => s.Contains("$group"));
-            group.ToJson().Should().Contain("$bottomN");
-            group.ToJson().Should().Contain("sortBy");
-        }
-
-        [Fact]
-        public void Mql_Constant_translates_under_custom_domain()
-        {
-            RequireServer.Check();
-
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            var client = DriverTestConfiguration.CreateMongoClient(c => c.SerializationDomain = customDomain);
-            var collection = client
-                .GetDatabase(DriverTestConfiguration.DatabaseNamespace.DatabaseName)
-                .GetCollection<EnumModel>(DriverTestConfiguration.CollectionNamespace.CollectionName);
-
-            var queryable = collection
-                .AsQueryable()
-                .Select(d => new { d.Id, V = d.ApprovalState == ApprovalState.Active ? Mql.Constant(ApprovalState.Active, BsonType.String) : ApprovalState.Inactive });
-
-            var stages = Linq3TestHelpers.Translate<EnumModel, object>(queryable);
-            var project = stages.Single(s => s.Contains("$project"));
-            project.ToJson().Should().Contain("Active");
-        }
+        var ex = Assert.Throws<ExpressionNotSupportedException>(() => Linq3TestHelpers.Translate<Person, Person>(queryable));
+        ex.Message.Should().EndWith("because Concat is not supported on queryables from different MongoClients (current domain 'outer', other queryable's domain 'inner').");
     }
+
+    [Fact]
+    public void Linq_Inject_translates_filter_under_outer_domain()
+    {
+        RequireServer.Check();
+
+        // Custom domain's CustomStringSerializer rewrites strings with a "test" suffix; the filter value
+        // must be serialized through the domain's registry when the Inject path is translated.
+        var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
+        customDomain.RegisterSerializer(new CustomStringSerializer());
+
+        var client = CreateClientWithDomain(customDomain, dropCollection: false);
+        var collection = GetTypedCollection<Person>(client);
+
+        var injectedFilter = Builders<Person>.Filter.Eq(p => p.Name, "Mario");
+        var queryable = collection.AsQueryable().Where(p => injectedFilter.Inject());
+
+        var stages = Linq3TestHelpers.Translate<Person, Person>(queryable);
+        var match = stages.Single(s => s.Contains("$match"));
+        match.ToJson().Should().Contain("Mariotest");
+    }
+
+    [Fact]
+    public void Linq_math_methods_translate_under_custom_domain()
+    {
+        RequireServer.Check();
+
+        // Register a renaming convention only on the custom domain so the rendered $project's
+        // field reference proves the Ceiling/Floor translators consulted this domain's class map.
+        var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
+        RegisterLowerCaseElementNameConvention(customDomain, typeof(NumericModel));
+
+        var client = CreateClientWithDomain(customDomain, dropCollection: false);
+        var collection = GetTypedCollection<NumericModel>(client);
+
+        var queryable = collection
+            .AsQueryable()
+            .Select(m => new { Ceil = System.Math.Ceiling(m.D), Flr = System.Math.Floor(m.D) });
+
+        var stages = Linq3TestHelpers.Translate<NumericModel, object>(queryable);
+        var project = stages.Single(s => s.Contains("$project"));
+        var projectJson = project.ToJson();
+        projectJson.Should().Contain("$ceil");
+        projectJson.Should().Contain("$floor");
+        // Lowercase 'd' (not 'D') only appears if NumericModel's class map was looked up via the custom domain.
+        projectJson.Should().Contain("$d");
+        projectJson.Should().NotContain("$D");
+    }
+
+    [Fact]
+    public void Linq_Pick_with_SortBy_runs_under_custom_domain()
+    {
+        RequireServer.Check();
+
+        // Register a renaming convention only on the custom domain so the rendered sortBy's
+        // field name proves Pick's SortDefinition render path consulted this domain.
+        var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
+        RegisterLowerCaseElementNameConvention(customDomain, typeof(Person));
+
+        var client = CreateClientWithDomain(customDomain, dropCollection: false);
+        var collection = GetTypedCollection<Person>(client);
+
+        var queryable = collection
+            .AsQueryable()
+            .GroupBy(p => p.Name)
+            .Select(g => new { Name = g.Key, YoungestAge = g.BottomN(Builders<Person>.Sort.Descending(p => p.Age), p => p.Age, 1) });
+
+        var stages = Linq3TestHelpers.Translate<Person, object>(queryable);
+        var group = stages.Single(s => s.Contains("$group"));
+        var groupJson = group.ToJson();
+        groupJson.Should().Contain("$bottomN");
+        groupJson.Should().Contain("sortBy");
+        // Lowercase 'age' (not 'Age') only appears if Person's class map was looked up via the custom domain.
+        groupJson.Should().Contain("\"age\" : -1");
+        groupJson.Should().NotContain("\"Age\"");
+    }
+
+    [Fact]
+    public void Mql_Constant_translates_under_custom_domain()
+    {
+        RequireServer.Check();
+
+        // Register a renaming convention only on the custom domain so the rendered $project's
+        // field reference proves Mql.Constant's translator consulted this domain's class map.
+        var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
+        RegisterLowerCaseElementNameConvention(customDomain, typeof(EnumModel));
+
+        var client = CreateClientWithDomain(customDomain, dropCollection: false);
+        var collection = GetTypedCollection<EnumModel>(client);
+
+        var queryable = collection
+            .AsQueryable()
+            .Select(d => new { d.Id, V = d.ApprovalState == ApprovalState.Active ? Mql.Constant(ApprovalState.Active, BsonType.String) : ApprovalState.Inactive });
+
+        var stages = Linq3TestHelpers.Translate<EnumModel, object>(queryable);
+        var project = stages.Single(s => s.Contains("$project"));
+        var projectJson = project.ToJson();
+        projectJson.Should().Contain("Active");
+        // Lowercase '$approvalstate' (not '$ApprovalState') only appears if EnumModel's class map
+        // was looked up via the custom domain.
+        projectJson.Should().Contain("$approvalstate");
+        projectJson.Should().NotContain("$ApprovalState");
+    }
+
+    private static void RegisterLowerCaseElementNameConvention(IBsonSerializationDomain domain, Type targetType)
+    {
+        var pack = new ConventionPack(domain);
+        pack.AddMemberMapConvention("LowerCaseElementName", m => m.SetElementName(m.MemberName.ToLower()));
+        domain.ConventionRegistry.Register("lowercase", pack, t => t == targetType);
+    }
+}
+
+internal enum ApprovalState { Active = 1, Inactive = 2 }
+
+internal class EnumModel
+{
+    [BsonId] public ObjectId Id { get; set; }
+    [BsonRepresentation(BsonType.String)]
+    public ApprovalState ApprovalState { get; set; }
 }
